+++ platforms/fabric-26.2/src/main/java/dev/cipherchannels/ui/ModMenuIntegration.java
@@ -1,0 +1,11 @@
+package dev.cipherchannels.ui;
+
+import com.terraformersmc.modmenu.api.ConfigScreenFactory;
+import com.terraformersmc.modmenu.api.ModMenuApi;
+
+public final class ModMenuIntegration implements ModMenuApi {
+    @Override
+    public ConfigScreenFactory<?> getModConfigScreenFactory() {
+        return ChannelManagerScreen::new;
+    }
+}